4em commit

+++ src/main/java/com/senforage/repos/VillageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.senforage.entities.Village;
 
-public interface VillageRepository extends JpaRepository<Village, Long>{
+public interface VillageRepository extends JpaRepository<Village, Integer>{
 
 }
